// src/components/Tabs/ChatModel/ChatModel.interface.ts
export enum sortTitles {
	CHECK = 'CHECK',
	ID = 'ID',
	USERNAME = 'USERNAME',
	MASTERNAME = 'MASTERNAME',
	ISSUE = 'ISSUE',
	STATUS = 'STATUS',
	DATE_UPDATED = 'DATE_UPDATED',
	DATE_CREATED = 'DATE_CREATED',
}

export enum sortTitlesView {
	CHECK = ' ',
	ID = '№',
	USERNAME = 'Имя пользователя',
	MASTERNAME = 'Имя мастера',
	ISSUE = 'Проблема',
	STATUS = 'Статус',
	DATE_UPDATED = 'Дата обновления',
	DATE_CREATED = 'Дата создания',
}

export enum ChatStatus {
	PENDING = 'PENDING',
	ACCEPTED = 'ACCEPTED',
}

export enum ChatStatusView {
	PENDING = 'В ожидании',
	ACCEPTED = 'Принят',
}